refactor: rename rover_chassis_controller -> rover_controller
