+++ src/rover_bringup/config/rover_controller.yaml
@@ -10,7 +10,7 @@ controller_manager:
       type: joint_state_broadcaster/JointStateBroadcaster
 
     swerve_controller:
-      type: rover_chassis_controller/RoverSwerveController
+      type: rover_controller/RoverSwerveController
 
 swerve_controller:
   ros__parameters:
