+++ src/rover_bringup/package.xml
@@ -40,7 +40,7 @@
   <exec_depend>controller_manager</exec_depend>
   <exec_depend>robot_state_publisher</exec_depend>
   <exec_depend>joint_state_broadcaster</exec_depend>
-  <exec_depend>rover_chassis_controller</exec_depend>
+  <exec_depend>rover_controller</exec_depend>
 
   <!-- Interfaces -->
   <exec_depend>lifecycle_msgs</exec_depend>
