--- src/rover_chassis_controller/plugins.xml
@@ -1,7 +1,0 @@
-<library path="rover_chassis_controller">
-  <class name="rover_chassis_controller/RoverSwerveController"
-         type="rover_chassis_controller::RoverSwerveController"
-         base_class_type="controller_interface::ControllerInterface">
-    <description>4-wheel swerve drive controller</description>
-  </class>
-</library>
