+++ src/rover_controller/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>rover_chassis_controller</name>
+  <name>rover_controller</name>
   <version>0.0.0</version>
   <description>4-wheel swerve drive ros2_control controller plugin</description>
   <maintainer email="fito.pn@ucu.edu.ua">yurifi</maintainer>
