+++ src/rover_controller/plugins.xml
@@ -1,0 +1,7 @@
+<library path="rover_controller">
+  <class name="rover_controller/RoverSwerveController"
+         type="rover_controller::RoverSwerveController"
+         base_class_type="controller_interface::ControllerInterface">
+    <description>4-wheel swerve drive controller</description>
+  </class>
+</library>
