+++ src/rover_sim/parameters/ros2_control.yaml
@@ -11,7 +11,7 @@ controller_manager:
       type: joint_state_broadcaster/JointStateBroadcaster
 
     swerve_controller:
-      type: rover_chassis_controller/RoverSwerveController
+      type: rover_controller/RoverSwerveController
 
 swerve_controller:
   ros__parameters:
